+++ lib/components/persistent_tab_view_scaffold.dart
@@ -191,10 +191,14 @@ class _PersistentTabViewScaffoldState extends State<PersistentTabViewScaffold>
           axisAlignment: -1,
           child: Padding(
             padding: widget.margin,
-            child: MediaQuery.removePadding(
-              context: context,
-              // safespace should be ignored, so the bottom inset is removed before it could be applied by any safearea child (e.g. in DecoratedNavBar).
-              removeBottom: !widget.avoidBottomPadding,
+            child: MediaQuery(
+              data: MediaQuery.of(context).copyWith(
+                padding: !widget.avoidBottomPadding
+                    // safespace should be ignored, so the bottom inset is removed before it could be applied by any safearea child (e.g. in DecoratedNavBar).
+                    ? EdgeInsets.zero
+                    // The padding might have been consumed by the keyboard, so it is maintained here. Using maintainBottomViewPadding would require that in the DecoratedNavBar as well, but only if the bottom padding should not be avoided. So it is easier to just maintain the padding here.
+                    : MediaQuery.of(context).viewPadding,
+              ),
               child: SafeArea(
                 top: false,
                 right: false,