+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replace NavBar SlideTransition with SizeTransition to correctly report its size to the Main Scaffold
 - Pass on the source padding to the Main Scaffold Body if the navbar is hidden
 - Pass the source viewPadding if navbar is hidden and the new body padding of the Main Scaffold as a new viewPadding to the tabs to ensure better placement of FABs on individual screens
+- Prevent Navbar jumping when keyboard opens because the keyboard consumes the MediaQuery.padding.bottom
+- Prevent Navbar jumping when keyboard opens because the keyboard consumes the MediaQuery.padding.bottom
 
 ## [5.0.0] - 2024-03-13
 ### Added